Done with specifications

+++ vagrant/catalog/project.py
@@ -1,6 +1,7 @@
-from sqlalchemy import create_engine
+from sqlalchemy import create_engine, asc, desc
 from sqlalchemy.orm import sessionmaker
-from catalogDB import Base, Categories, Items
+from catalogDB import Base, Category, Item,User
+# , User
 #imported the Flask class
 from flask import Flask, render_template, request, redirect, url_for, flash, jsonify
 #An instance of this class is created which is our WSGI application.
@@ -35,7 +36,9 @@ import random, string
 # import string
 # Create an engine that stores data in the local directory's
 # sqlalchemy_example.db file.
-engine = create_engine('sqlite:///catalog_app.db')
+# engine = create_engine('sqlite:///catalog_app.db')
+engine = create_engine('sqlite:///groceryCatalog3.db')
+# engine = create_engine('sqlite:///catalog_appWithUsers2.db')
 # Create all tables in the engine. This is equivalent to "Create Table"
 # statements in raw SQL.
 Base.metadata.bind = engine
@@ -117,13 +120,18 @@ def gconnect():
     stored_credentials = login_session.get('credentials')
     stored_gplus_id = login_session.get('gplus_id')
     if stored_credentials is not None and gplus_id == stored_gplus_id:
+        # Store the access token in the session for later use.
+        del login_session['access_token']
+        login_session['access_token'] = credentials.access_token
+        login_session['gplus_id'] = gplus_id
         response = make_response(json.dumps('Current user is already connected.'),
                                  200)
         response.headers['Content-Type'] = 'application/json'
         return response
 
     # Store the access token in the session for later use.
-    login_session['credentials'] = credentials.access_token
+    # login_session['credentials'] = credentials.access_token
+    login_session['access_token'] = credentials.access_token
     login_session['gplus_id'] = gplus_id
 
     # Get user info
@@ -137,6 +145,12 @@ def gconnect():
     login_session['picture'] = data['picture']
     login_session['email'] = data['email']
 
+    # see if user exists, if it doesn't make a new one
+    user_id = getUserID(login_session['email'])
+    if not user_id:
+        user_id = createUser(login_session)
+    login_session['user_id'] = user_id
+
     output = ''
     output += '<h1>Welcome, '
     output += login_session['username']
@@ -144,15 +158,38 @@ def gconnect():
     output += '<img src="'
     output += login_session['picture']
     output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
-    flash("you are now logged in as %s" % login_session['username'])
+    flash("you are now logged in as %s" % login_session['username'], 'login')
     print "done!"
     return output
 
+
+def createUser(login_session):
+    newUser = User(name=login_session['username'], email=login_session[
+                   'email'], picture=login_session['picture'])
+    session.add(newUser)
+    session.commit()
+    user = session.query(User).filter_by(email=login_session['email']).one()
+    return user.id
+
+
+def getUserInfo(user_id):
+    user = session.query(User).filter_by(id=user_id).one()
+    return user
+
+
+def getUserID(email):
+    try:
+        user = session.query(User).filter_by(email=email).one()
+        return user.id
+    except:
+        return None
+
 # DISCONNECT - Revoke a current user's token and reset their login_session
 # disconnect by telling the server to reject its access token
-@app.route('/gdisconnect')
+@app.route('/logout')
 def gdisconnect():
-    access_token = login_session['access_token']
+    print(login_session)
+    access_token = login_session.get('access_token')
     print 'In gdisconnect access token is %s', access_token
     print 'User name is: '
     print login_session['username']
@@ -161,13 +198,14 @@ def gdisconnect():
     	response = make_response(json.dumps('Current user not connected.'), 401)
     	response.headers['Content-Type'] = 'application/json'
     	return response
-    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % login_session['access_token']
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % login_session.get('access_token')
     h = httplib2.Http()
     result = h.request(url, 'GET')[0]
     print 'result is '
     print result
     if result['status'] == '200':
-        del login_session['access_token']
+        access_token = login_session.get('access_token')
+        del access_token
     	del login_session['gplus_id']
     	del login_session['username']
     	del login_session['email']
@@ -180,62 +218,117 @@ def gdisconnect():
     	response = make_response(json.dumps('Failed to revoke token for given user.', 400))
     	response.headers['Content-Type'] = 'application/json'
     	return response
-#route() decorator to tell Flask what URL should trigger our function
-@app.route('/')
-def HelloWorld():
-	return 'Hello,World!'
-@app.route('/category/<int:category_id>/')
-def HelloCatalog(category_id):
-	category = session.query(Categories).filter_by(id=category_id)
-	items = session.query(Items).filter_by(category_id = category_id)
-	return render_template('menu.html',category=category[0], items=items)
-
 
 @app.route('/category/<int:category_id>/item/JSON')
 def categoriesItemJSON(category_id):
-    category = session.query(Categories).filter_by(id=category_id).one()
-    items = session.query(Items).filter_by(
+    category = session.query(Category).filter_by(id=category_id).one()
+    items = session.query(Item).filter_by(
         category_id=category.id).all()
     return jsonify(Items=[i.serialize for i in items])
 
 # ADD JSON ENDPOINT HERE
 @app.route('/category/<int:category_id>/item/<int:item_id>/JSON')
 def menuItemJSON(category_id, item_id):
-    item = session.query(Items).filter_by(id=item_id).one()
+    item = session.query(Item).filter_by(id=item_id).one()
     return jsonify(Item=item.serialize)
+
+#route() decorator to tell Flask what URL should trigger our function
+@app.route('/')
+def HelloWorld():
+    return render_template('home.html')
+
+
+@app.route('/category/<int:category_id>/')
+def showACategory(category_id):
+    category = session.query(Category).filter_by(id=category_id)
+    items = session.query(Item).filter_by(category_id = category_id)
+    return render_template('item.html',category=category[0], items=items)
+
+
+# Show all restaurants
+# @app.route('/')
+@app.route('/category/')
+def showCategories():
+    categories = session.query(Category).order_by(asc(Category.id)).all()
+    items = session.query(Item).group_by(Item.category_id).order_by(desc(Item.created_date)).all()
+    print(items)
+    for category in categories:
+        # print(category.id)
+        for item in items:
+            if item.category_id == category.id:
+                category.item = item
+                break
+    if 'username' not in login_session:
+        return render_template('publiccategories.html', categories=categories)
+    else:
+        return render_template('categories.html', categories=categories)
+
+# Create a new category
+@app.route('/category/new/', methods=['GET', 'POST'])
+def newCategory():
+    if 'username' not in login_session:
+        return redirect('/login')
+    if request.method == 'POST':
+        newCategory = Category(
+            name=request.form['name'], description = request.form['description'],
+            user_id=login_session.get('user_id')
+            )
+        session.add(newCategory)
+        flash('New Category %s Successfully Created' % newCategory.name, 'newCategory')
+        session.commit()
+        return redirect(url_for('showCategories'))
+    else:
+        return render_template('newCategory.html')
+
+
+
 
 @app.route('/category/<int:category_id>/new/', methods=['GET','POST'])
 def newItem(category_id):
     if 'username' not in login_session:
         return redirect('/login')
-	if request.method == 'POST':
-		newItem = Items(name=request.form['name'],
-			description=request.form['description'],
-			category_id=category_id)
-		session.add(newItem)
-		session.commit()
-		flash("new menu item created!")
-		return redirect(url_for('HelloCatalog',category_id=category_id))
-	else:
-		return render_template('newItem.html', category_id=category_id)
+    newItemInCate = session.query(Category).filter_by(id = category_id).one()
+    if login_session['user_id'] == newItemInCate.user_id:
+        if request.method == 'POST':
+            print("I am here!")
+            newItem = Item(name=request.form['name'],
+    			description=request.form['description'],
+    			category_id=category_id,
+                user_id=login_session.get('user_id')
+                )
+            print(newItem)
+            session.add(newItem)
+            session.commit()
+            flash("new item created!", 'newItem')
+            return redirect(url_for('showACategory',category_id=category_id))
+        else:
+            return render_template('newItem.html', category_id=category_id)
+    else:
+        flash("You are not authorized to add an item in the selected category!",'notAuthorized')
+        return redirect(url_for('showACategory',category_id=category_id))
 
 @app.route('/category/<int:category_id>/<int:item_id>/edit/', methods=['GET','POST'])
 def editItem(category_id, item_id):
     if 'username' not in login_session:
         return redirect('/login')
-    editedItem = session.query(Items).filter_by(id = item_id).one()
-    if request.method == 'POST':
-		if request.form['name']:
-			editedItem.name = request.form['name']
-			editedItem.description = request.form['description']
-		session.add(editedItem)
-		session.commit()
-		flash("The selected item is edited!")
-		return redirect(url_for('HelloCatalog',category_id=category_id))
+    editedItem = session.query(Item).filter_by(id = item_id).one()
+    if login_session['user_id'] == editedItem.user_id:
+        if request.method == 'POST':
+            if request.form['name']:
+                editedItem.name = request.form['name']
+                editedItem.description = request.form['description']
+                session.add(editedItem)
+                session.commit()
+                flash("The selected item is edited!","editItem")
+                return redirect(url_for('showACategory',category_id=category_id))
+            if not request.form['name']:
+                return redirect(url_for('showACategory',category_id=category_id))
+        else:
+            return render_template('editItem.html', category_id=category_id,
+                item_id = item_id,item = editedItem)
     else:
-		return render_template('editItem.html', category_id=category_id,
-			item_id = item_id,item = editedItem)
-	# return "page to edit a menu item. Task 2 complete!"
+        flash("You are not authorized to edit the selected item!",'notAuthorized')
+        return redirect(url_for('showACategory',category_id=category_id))
 
 
 
@@ -243,15 +336,19 @@ def editItem(category_id, item_id):
 def deleteItem(category_id, item_id):
     if 'username' not in login_session:
         return redirect('/login')
-    deletedItem = session.query(Items).filter_by(id = item_id).one()
-    if request.method == 'POST':
-		session.delete(deletedItem)
-		session.commit()
-		flash("The selected Item  is deleted!")
-		return redirect(url_for('HelloCatalog',category_id=category_id))
+    deletedItem = session.query(Item).filter_by(id = item_id).one()
+    if login_session['user_id'] == deletedItem.user_id:
+        if request.method == 'POST':
+            session.delete(deletedItem)
+            session.commit()
+            flash("The selected Item  is deleted!", "deleteItem")
+            return redirect(url_for('showACategory',category_id=category_id))
+        else:
+            return render_template('deleteItem.html', category_id=category_id,
+            item_id = item_id,item = deletedItem)
     else:
-		return render_template('deleteItem.html', category_id=category_id,
-			item_id = item_id,item = deletedItem)
+        flash("You are not authorized to delete the selected item!",'notAuthorized')
+        return redirect(url_for('showACategory',category_id=category_id))
 
 if __name__ == '__main__':
 	app.secret_key = 'super_secret_key'
@@ -260,4 +357,4 @@ if __name__ == '__main__':
 	app.debug = True
 	# This will immediately launch a local server and
 	# start the server on http://localhost:5000/.
-	app.run(host='0.0.0.0', port = 5000)
+	app.run(host='0.0.0.0', port = 5080)
